LiveGrip/app/requirements.txt
django
djangorestframework
gunicorn
psycopg2
rethinkdb
requests
argon2-cffi
